extracted ScriptProjectProvider
